docs(aesutils): document utility implementation
Add descriptive comment in AESUtils.cpp.
